+++ packages/css-color-parser/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add `computedValue()` serialization function for color data.
 - Mandatory `alpha` in the relative `alpha()` function.
+- Updated [`@csstools/color-helpers`](https://github.com/csstools/postcss-plugins/tree/main/packages/color-helpers) to [`6.1.1`](https://github.com/csstools/postcss-plugins/tree/main/packages/color-helpers/CHANGELOG.md#611) (patch)
 
 ### 4.1.10